% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf.R
\name{abf.fit.parallel.setup}
\alias{abf.fit.parallel.setup}
\title{Internal function: create parallel.dir, save necessary files}
\usage{
abf.fit.parallel.setup(parallel.dir, snps, x2, y2, q, family)
}
\arguments{
\item{parallel.dir}{directory to create}

\item{snps}{snps to model}

\item{x2}{genotype data}

\item{y2}{phenotype data}

\item{q}{covariates}

\item{family}{glm family}

\item{...}{items to save in data.RData under parallel}
}
\value{
No return value
}
\description{
Internal function: create parallel.dir, save necessary files
}
\author{
chris
}
