% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf.R
\name{abf.speedglm.fit}
\alias{abf.speedglm.fit}
\title{abf.speedglm.fit}
\usage{
abf.speedglm.fit(x, y, q, family, snps, parallel.dir = NULL,
  verbose = FALSE)
}
\arguments{
\item{x}{design matrix}

\item{y}{response vector}

\item{q}{matrix of covariates}

\item{family}{glm family}

\item{snps}{list of snps models}

\item{parallel.dir}{optional directory used for manual parallelisation}

\item{verbose}{if TRUE, print progress. Default is FALSE}
}
\value{
a list of bics and coeffs, or number of models to be fit if parallel.dir is not NULL
}
\description{
Internal function for fitting abf
}
\author{
Chris Wallace
}
