% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{snps}
\alias{snps}
\alias{snps,groups-method}
\alias{snps,tags-method}
\alias{tags}
\alias{tags,groups-method}
\alias{tags,tags-method}
\title{Accessors for groups objects}
\usage{
snps(object)

tags(object)

\S4method{snps}{groups}(object)

\S4method{tags}{groups}(object)

\S4method{snps}{tags}(object)

\S4method{tags}{tags}(object)
}
\arguments{
\item{object}{object from which items should be extracted}
}
\value{
a list of character vectors
}
\description{
Accessor functions
}
\details{
\code{snps} shows list of snps in an object of class groups and
returns a list of character vectors

\code{tags} shows tags from an object of class groups and
returns a character vector of tag SNPs
}
\author{
Chris Wallace
}

