\name{addlines}
\alias{addlines}
\title{addlines}
\usage{
addlines(summ, alpha = 0.5, size = 0.2, xvar = "position.plot")
}
\arguments{
  \item{summ}{data.frame generated by guess.summ}

  \item{alpha}{transparency, default=0.5}

  \item{size}{line width, default=0.2}
}
\value{
object that can be added to a ggplot
}
\description{
Add lines to a ggplot, indicating position of snps
}
\author{
Chris Wallace
}

