% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{best.models}
\alias{best.models}
\title{Best models}
\usage{
best.models(d, pp.thr = 0.01, cpp.thr = NA)
}
\arguments{
\item{d}{snpmod object or a list of snpmod objects}

\item{pp.thr}{models with PP>pp.thr will be shown}

\item{cpp.thr}{the smallest set of models such that sum(PP)>=cpp.thr will be shown}
}
\value{
subset of \code{models} slot in \code{d} showing best models
}
\description{
Display most likely models according to their posterior probability
}
\details{
Note that this isn't posterior probability in the true sense, it is the posterior probability for each model amongst the bag of models in the snpmod object, \code{d}.
}
\author{
chris
}
