% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{calc.maxmin}
\alias{calc.maxmin}
\title{calculate max or min of subset of a matrix}
\usage{
calc.maxmin(r2, L1, L2 = L1, fun = max)
}
\arguments{
\item{r2}{matrix, with row/column names which are members of L1, L2}

\item{L1}{first list of vectors of snps}

\item{L2}{second list vectors of snps}

\item{fun}{max or min, but could be another function that returns a scalar. unquoted.}
}
\value{
smaller matrix indexed by elements of L1 (rows) and L2 (columns)
}
\description{
calculate max or min of subset of a matrix
}
\author{
Chris Wallace
}
