% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{cond.best}
\alias{cond.best}
\title{cond.best}
\usage{
cond.best(X, Y, best = NULL, stepwise.p.thr = 0.001,
  stepwise.max.predictors = NA, ...)
}
\arguments{
\item{X}{snpMatrix object}

\item{Y}{phenotype vector}

\item{best}{SNPs chosen so far, NULL if none}

\item{stepwise.p.thr}{maximum p value to continue adding predictors}

\item{stepwise.max.predictors}{maximum predictors to choose, NA if continue until stepwise.p.thr is met}

\item{...}{arguments passed to snp.rhs.test}
}
\value{
new predictor, NULL if none found that meets conditions
}
\description{
Stepwise regression on a snpMatrix
}
\details{
Calls snp.rhs.test to find next best predictor to add, possibly given a list of existing chosen predictors.
}
\author{
Chris Wallace
}
