% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{convert,groups-method}
\alias{convert,groups-method}
\alias{convert,tags-method}
\title{Conversion of objects saved in pre-release version of GUESSFM to
current version}
\usage{
\S4method{convert}{groups}(object)

\S4method{convert}{tags}(object)
}
\arguments{
\item{object}{object of class groups or tags}
}
\value{
object of the same class
}
\description{
Conversion of objects saved in pre-release version of GUESSFM to
current version
}

