% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert}
\alias{convert}
\title{Convert from old definitions of groups, tags classes to new
DON'T USE THIS FUNCTION UNLESS YOU HAVE OBJECTS STORED FROM A PREVIOUS PRE-DEVELOPMENT VERSION!}
\usage{
convert(object)
}
\arguments{
\item{object}{GUESSFM object from pre-development version}
}
\value{
new S4 structure
}
\description{
Convert from old definitions of groups, tags classes to new
DON'T USE THIS FUNCTION UNLESS YOU HAVE OBJECTS STORED FROM A PREVIOUS PRE-DEVELOPMENT VERSION!
}

