% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{cummean}
\alias{cummean}
\title{internal function: cumulative mean}
\usage{
cummean(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
vector of cumulative mean
}
\description{
internal function: cumulative mean
}
\examples{
GUESSFM:::cummean(1:10)
}

