% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{expand.tags}
\alias{expand.tags}
\title{Expand tags for a snpmod object}
\usage{
expand.tags(d, tags)
}
\arguments{
\item{d}{snpmod object}

\item{tags}{tags object}
}
\value{
snpmod object with tags expanded
}
\description{
We can use tagging to reduce the number of models.  This function expands each tagged model to every possible model that could be created from its component (tag) SNPs.
}

