% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{ggbed}
\alias{ggbed}
\title{Plot a bed file}
\usage{
ggbed(bed, summ)
}
\arguments{
\item{bed}{data.frame representing data read in from a bed file}

\item{summ}{data.frame generated by \code{\link{guess.summ}()}}
}
\value{
ggplot object
}
\description{
Plot a bed file, overlaying lines showing the positions of SNPs in a summ object
}
\author{
chris
}
\seealso{
Other plotting.GUESSFM.results: \code{\link{addlines}};
  \code{\link{ggchr}}; \code{\link{ggld}};
  \code{\link{pp.nsnp}}; \code{\link{scalepos}};
  \code{\link{signal.plot}}
}

