% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{ggchr}
\alias{ggchr}
\title{ggchr}
\usage{
ggchr(summ)
}
\arguments{
\item{summ}{data.frame generated by \code{\link{guess.summ}()}}
}
\value{
ggplot2 object
}
\description{
Plot lines connecting each SNP's physical position to their representative position in the summ object
}
\author{
chris
}
\seealso{
Other plotting.GUESSFM.results: \code{\link{addlines}};
  \code{\link{ggbed}}; \code{\link{ggld}};
  \code{\link{pp.nsnp}}; \code{\link{scalepos}};
  \code{\link{signal.plot}}
}

