% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{ggld}
\alias{ggld}
\title{ggld}
\usage{
ggld(data, df.snps)
}
\arguments{
\item{data}{object of class SnpMatrix from which LD will be calculated}

\item{df.snps}{data.frame generated by guess.summ}
}
\value{
ggplot output
}
\description{
Rotated LD plot
}
\details{
Generates a plot of r2, rotating through 45 degrees so it can be
aligned with other results, similarly to Haploview.

The positions in the df.snps object are used to align the snps to
the correct coordinates.
}
\author{
Chris Wallace
}

