% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggsnp}
\alias{ggsnp}
\title{ggsnp}
\usage{
ggsnp(summx, snp.name = "snp")
}
\arguments{
\item{summx}{summary of a snpmod + groups object}

\item{snp.name}{column name for snp labels, default \code{snp}}
}
\value{
a ggplot object
}
\description{
create a track from a guess.summ object showing snp ids
}
\author{
Chris Wallace
}
