% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group.multi}
\alias{group.multi}
\title{Group SNPs}
\usage{
group.multi(SM2, snp.data, min.mppi = 0.01, r2.minmerge = 0.5)
}
\arguments{
\item{SM2}{snpmod object}

\item{snp.data}{SnpMatrix for r2 calculation}

\item{min.mppi}{trim snp groups with total MPPI < min.mppi in all
diseases}

\item{r2.minmerge}{merge groups with minimum between-group r2 >
r2.minmerge}
}
\value{
list with three components.

First is a data.frame with each row giving summary statistics for
    each group.

Second is a groups object, each elements ordered according to the rows of the summary

Third is the r2 matrix calculated.
}
\author{
Chris Wallace
}
