\docType{class}
\name{groups-class}
\alias{groups-class}
\title{groups class}
\description{
groups class
}
\section{Slots}{

\describe{
\item{\code{tags}}{character vector giving tag SNPs.  Each tag indexes one group of SNPs}

\item{\code{groups}}{list of character vectors giving the SNP membership of each group}
}}
\section{Slots}{

}

