% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{methods}
\name{[,groups,character,missing,missing-method}
\alias{[,groups,character,missing,missing-method}
\alias{[,groups,logical,missing,missing-method}
\alias{[,groups,numeric,missing,missing-method}
\alias{[,tags,character,missing,missing-method}
\alias{[[,groups,character-method}
\alias{[[,groups,logical-method}
\alias{[[,groups,numeric-method}
\title{Subset groups or tags objects}
\usage{
\S4method{[}{groups,character,missing,missing}(x, i)

\S4method{[}{groups,numeric,missing,missing}(x, i)

\S4method{[}{groups,logical,missing,missing}(x, i)

\S4method{[}{tags,character,missing,missing}(x, i)

\S4method{[[}{groups,numeric}(x, i)

\S4method{[[}{groups,logical}(x, i)

\S4method{[[}{groups,character}(x, i)
}
\arguments{
\item{x}{groups object}

\item{i}{numeric, logical or character vector to index SNPs}
}
\value{
subsetted groups or tags object
}
\description{
'[' will extract another object of the same class.  '[[' will extract a single element.
}

