% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{guess.summ}
\alias{guess.summ}
\title{Summarise a snpmod object}
\usage{
guess.summ(results, groups = NULL, snps = NULL, snp.data = NULL,
  tag.thr = 0.8, pp.thr = 0.01, method = "complete")
}
\arguments{
\item{results}{an object of class snpmod, or a named list of such objects}

\item{groups}{object of class groups.  If supplied, all SNPs here will be summarised, and grouped according to this structure}

\item{snps}{data.frame giving details of the SNPs}

\item{snp.data}{if groups is missing, tag groups are determined from this object of class SnpMatrix}

\item{tag.thr}{if groups is missing, threshold at which to tag}

\item{pp.thr}{if groups is missing, threshold above which SNPs are selected for summary.}

\item{method}{if groups is missing, method to determine tag groups using heirarchical clustering, default is "complete"}
}
\value{
data.frame
}
\description{
guess.summ
}
\author{
Chris Wallace
}

