% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makestr}
\alias{makestr}
\title{Paste a list of SNPs in a consistent way to create a model name}
\usage{
makestr(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character object
}
\description{
Paste a list of SNPs in a consistent way to create a model name
}
\examples{
SNPS1 <- sample(letters[1:10])
SNPS2 <- sample(letters[1:10])
SNPS1
SNPS2
GUESSFM:::makestr(SNPS1)
GUESSFM:::makestr(SNPS2)
}
