% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-union.R
\name{model.union}
\alias{model.union}
\title{Model unions}
\usage{
model.union(e1, e2, dataset.labels = c(1, 2), detail = FALSE)
}
\arguments{
\item{e1}{snpmod object 1 or a character vectors of model strings extracted from a snpmod}

\item{e2}{snpmod object 2 or a character vectors of model strings extracted from a snpmod}

\item{dataset.labels}{labels used to refer to e1 and e2 in a detailed union of models. Ignored unless \code{detail==TRUE}.}

\item{detail}{default FALSE. If TRUE, returns a more detailed data.frame showing union of models and from which dataset they came.}
}
\value{
If \code{detail=FALSE}, invisibly returns a character vector of the unique model strings found in e1 and e2.  If \code{detail=TRUE}, invisibly returns a data.frame containing that vector and some information on which dataset each was found in.
}
\description{
Describe and extract the union of models contained in two snpmod objects
}
\author{
Chris Wallace
}
