% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{nmodes}
\alias{nmodes}
\title{nmodes}
\usage{
nmodes(x)
}
\arguments{
\item{x}{pp.nsnp object}
}
\value{
list of numerics
}
\description{
Summarise number of modes of a posterior distribution on the number of SNPs in a model
}
\author{
Chris Wallace
}
