% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{overlap}
\alias{overlap}
\title{overlap}
\usage{
overlap(x, y)
}
\arguments{
\item{x}{object of class \code{groups}, \code{snppicker} or \code{tags}}

\item{y}{object of same class as x}
}
\value{
object of class groups
}
\description{
overlap can be used to examine the overlap between groups before the union is created
}
\author{
Chris Wallace
}
