% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pattern.plot}
\alias{pattern.plot}
\title{pattern.plot}
\usage{
pattern.plot(SM, groups, r2 = NULL)
}
\arguments{
\item{SM}{snpmod or list of snpmods}

\item{groups}{groups object}

\item{r2}{optional matrix of r2 between SNPs. If provided, will be
used to calculate maximum r2 between groups, which will be
displayed in the plot}
}
\value{
a ggplot object, by default printed to current graphics
  device
}
\description{
Plot pattern of SNP group inclusion
}
\author{
Chris Wallace
}
