% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{methods}
\name{plot,snppicker,missing-method}
\alias{plot,snppicker,missing-method}
\title{Plots}
\usage{
\S4method{plot}{snppicker,missing}(x, do.plot = TRUE)
}
\arguments{
\item{x}{snppicker object}

\item{do.plot}{if TRUE (default) print the plot on the current
device, otherwise invisibly return the plot object.}
}
\description{
Descriptive plots showing how the sets in snppicker objects were
generated.  Uses \code{ggplot2}, so you can customize using
\code{theme} etc.
}

