% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotsummary}
\alias{plotsummary}
\title{Summary plots}
\usage{
plotsummary(results)
}
\arguments{
\item{results}{object of class snpmod, or a named list of such objects}
}
\value{
a list of two ggplot2 objects, which may be print()ed to the current graphics device
}
\description{
Plot two summaries of the models, a diffusion plot and a summary of the prior and posterior number of SNPs in the models
}
