% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{pp.nsnp}
\alias{pp.nsnp}
\title{Summarize the posterior model support by the number of SNPs contained in a model}
\usage{
pp.nsnp(results, plot = FALSE)
}
\arguments{
\item{results}{object of class snpmod}

\item{plot}{if TRUE, print a pretty plot to current plotting device}
}
\value{
a list containing a named vector summarizing the posterior for each count of SNPs and a ggplot object
}
\description{
Summarize the posterior model support by the number of SNPs contained in a model
}
\seealso{
Other plotting.GUESSFM.results: \code{\link{addlines}};
  \code{\link{ggbed}}; \code{\link{ggchr}};
  \code{\link{ggld}}; \code{\link{scalepos}};
  \code{\link{signal.plot}}
}

