% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ppnsnp-class}
\alias{ppnsnp-class}
\title{Class to hold results of pp.nsnp}
\description{
\code{pp.nsnp} summarises prior and posterior support for the
number of SNPs required to model a trait or several traits
}
\section{Slots}{

\describe{
\item{\code{.Data}}{contains a named list of numeric vectors giving the support for each number}

\item{\code{plot}}{contains a ggplot object}

\item{\code{traits}}{names of traits, corresponding to items in .Data}
}}

\examples{
new("ppnsnp")
}
