% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/junkyard.R
\name{run.bvs.v1}
\alias{run.bvs.v1}
\title{Bayesian variable selection}
\usage{
run.bvs.v1(X, Y, gdir = "test", sub = NA, covars = NULL,
  family = "gaussian", nsweep = 55000, nchains = 3, nexp = 3,
  boot = 0, tag.r2 = 0.99, nsave = 1000, as.is = TRUE,
  dominance = FALSE, guess.command = "GUESS", backend = c("guess",
  "dummy"))
}
\arguments{
\item{X}{snpMatrix object, holding genotypes}

\item{Y}{vector or matrix of phenotypes, nrow(Y)==nrow(X)}

\item{gdir}{directory where all the GUESS input/output files will be. If it doesn't exist, it will be created.}

\item{sub}{optional number < nrow(X). If supplied, only the subset of samples defined by 1:sub will be used.}

\item{covars}{optional matrix or vector of covariates. If supplied GUESS will be run on residuals from glm(Y ~ ., data=as.data.frame(covars)).}

\item{family}{family for Y ~ covars regression.  default "gaussian".}

\item{nsweep}{number of sweeps, see documentation for GUESS}

\item{nchains}{number of chains, see documentation for GUESS}

\item{nexp}{expected number of causal variants in region}

\item{boot}{EXPERIMENTAL. Optional integer > 0 indicates that a number of bootstrap replicates should be run to understand variability in output}

\item{tag.r2}{r squared value at which to tag to avoid numerical instability.  Default of 0.99 has worked well in our experience.}

\item{nsave}{number of models to save, see documentation for GUESS}

\item{as.is}{EXPERIMENTAL if boot=TRUE, and as.is=TRUE, explicitly exclude the unchanged data structure from the set of bootstrapped datasets}

\item{dominance}{EXPERIMENTAL. If TRUE, include a dominance effect at each SNP}

\item{guess.command}{Command to run GUESS, if GUESS is not on your PATH.}

\item{backend}{"GUESS" (default) to run GUESS, or "dummy" to just generate the files for checking}
}
\value{
nothing.  side effect is to set GUESS running in the background.  This takes a while (typically several hours).
}
\description{
Wrapper for GUESS, an alternative to using R2GUESS
}

