% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{show,snpmod-method}
\alias{show,groups-method}
\alias{show,ppnsnp-method}
\alias{show,snpmod-method}
\alias{show,snppicker-method}
\alias{show,tags-method}
\title{Show}
\usage{
\S4method{show}{snpmod}(object)

\S4method{show}{snppicker}(object)

\S4method{show}{ppnsnp}(object)

\S4method{show}{tags}(object)

\S4method{show}{groups}(object)
}
\arguments{
\item{object}{the thing to show}
}
\description{
Methods to briefly summarise objects to screen. \code{show} is
called implicitly when you type an object's name at the command
line.  Use \code{summary}, where available, to get more details.
}

