% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{snpin}
\alias{snpin}
\alias{snpin,character,snppicker-method}
\alias{snpin,character,tags-method}
\alias{snpin,character,groups-method}
\title{Check whether a snp is in a snppicker, groups or tags object}
\usage{
snpin(x, y)

\S4method{snpin}{character,snppicker}(x, y)

\S4method{snpin}{character,tags}(x, y)

\S4method{snpin}{character,groups}(x, y)
}
\arguments{
\item{x}{character vector of SNPs to be checked}

\item{y}{object of class snppicker, groups or tags}
}
\value{
logical matrix of nrow equal to length(x) and ncol equal to number of groups in y
}
\description{
Check whether a snp is in a snppicker, groups or tags object
}
\author{
Chris Wallace
}
