\docType{class}
\name{snpmod-class}
\alias{snpmod-class}
\title{Class to hold data relating to multiple models fitted to SNP data}
\description{
Class to hold data relating to multiple models fitted to
SNP data
}
\section{Slots}{

\describe{
\item{\code{snps}}{data.frame containing marginal probabilities of inclusion for the SNPs}

\item{\code{models}}{data.frame containing summaries for each model}

\item{\code{model.snps}}{list containing the SNPs for each model. May be removed.}
}}
\section{Slots}{

}

