% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{snppicker-class}
\alias{show,snppicker-method}
\alias{snppicker-class}
\title{Class to hold results of snp.picker algorithm}
\usage{
\S4method{show}{snppicker}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
\code{snp.picker} groups SNPs according to LD and model inclusion
and outputs objects of class snppicker.
}
\section{Slots}{

\describe{
\item{\code{groups}}{list of data.frames describing SNPs in each group}

\item{\code{plotsdata}}{list of additional data relating to the snp.picker
process that allows a summary of that process to be plotted via
\code{plot}.}
}}

