% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{snps.from.correlated.models}
\alias{snps.from.correlated.models}
\title{snps.from.correlated.models}
\usage{
snps.from.correlated.models(ret, thr = 0.8, nshow = 10)
}
\arguments{
\item{ret}{data.frame generated by qc() on a snpmod}

\item{thr}{r2 threshold considered unlikely}

\item{nshow}{number of most frequently occurring SNPs to show}
}
\value{
the nshow top SNPs, if any
}
\description{
internal function: summarise snps from highly correlated models.
}
\author{
Chris Wallace
}
