% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{snps}
\alias{snps}
\alias{snps,groups-method}
\alias{snps,tags-method}
\title{snps}
\usage{
snps(object)

\S4method{snps}{groups}(object)

\S4method{snps}{tags}(object)
}
\arguments{
\item{object}{object from which snps are to be extracted}
}
\value{
a list of character vectors
}
\description{
Accessor function to show list of snps in an object of class groups
}
\author{
Chris Wallace
}

