% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf.R
\name{speedglm.wfit1}
\alias{speedglm.wfit1}
\title{speedglm.wfit1}
\usage{
speedglm.wfit1(y, X, intercept = TRUE, weights = NULL, row.chunk = NULL,
  family = gaussian(), start = NULL, etastart = NULL, mustart = NULL,
  offset = NULL, acc = 1e-08, maxit = 25, k = 2, sparselim = 0.9,
  camp = 0.01, eigendec = TRUE, tol.values = 1e-07, tol.vectors = 1e-07,
  tol.solve = .Machine$double.eps, sparse = NULL, method = c("eigen",
  "Cholesky", "qr"), trace = FALSE, ...)
}
\description{
Internal adaption of speedglm function speedglm.wfit1 due to dependency issue.  See speeglm package for thorough documentation.
}
\author{
Jamie Inshaw
}

