% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summ.setminmax}
\alias{summ.setminmax}
\title{summ.setminmax}
\usage{
summ.setminmax(df.snps)
}
\arguments{
\item{df.snps}{guess.summ object}
}
\value{
df.snps
}
\description{
set xmin, xmax on a guess.summ object for easy plotting
}
\details{
resets snpnum to be a continuous series of integers from 1
}
\author{
Chris Wallace
}
