% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{summary.plots}
\alias{summary.plots}
\title{Summary plots}
\usage{
summary.plots(results)
}
\arguments{
\item{results}{object of class snpmod, or a named list of such objects}
}
\value{
a list of two ggplot2 objects, which may be print()ed to the current graphics device
}
\description{
Plot two summaries of the models, a diffusion plot and a summary of the prior and posterior number of SNPs in the models
}

