% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{methods}
\name{summary,snppicker-method}
\alias{summary,groups-method}
\alias{summary,snppicker-method}
\title{Summaries}
\usage{
\S4method{summary}{snppicker}(object)

\S4method{summary}{groups}(object)
}
\arguments{
\item{object}{the thing to summarise}
}
\description{
Print summary of an object
}

