% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{tag}
\alias{tag}
\title{tag}
\usage{
tag(X, tag.threshold = 0.99, snps = NULL, samples = NULL, strata = NULL,
  quiet = FALSE, method = "complete")
}
\arguments{
\item{X}{SnpMatrix object}

\item{tag.threshold}{threshold to cut tree, default=0.99}

\item{snps}{colnames of the SnpMatrix object to be used}

\item{samples}{optional, subset of samples to use}

\item{strata}{optional, return a list of tag vectors, one for each stratum defined by as.factor(strata)}

\item{quiet}{if FALSE (default), show progress messages}

\item{method}{method used for heirarchical clustering.  See hclust for options.}
}
\value{
character vector, names are \code{snps}, values are the tag for each SNP
}
\description{
Derive tag SNPs for a SnpMatrix object using heirarchical clustering
}
\details{
Uses complete linkage and the \code{\link{hclust}} function to define clusters,
then cuts the tree at 1-tag.threshold
}
\author{
Chris Wallace
}

