\docType{class}
\name{tags-class}
\alias{tags-class}
\title{tags class}
\description{
tags class
}
\section{Slots}{

\describe{
\item{\code{tags}}{character vector giving tag SNPs, one per SNP in \code{snps}, repeated as necessary}

\item{\code{snps}}{character vector giving SNPs included in this tags object}
}}
\section{Slots}{

}

