% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{tags}
\alias{tags}
\alias{tags,groups-method}
\alias{tags,tags-method}
\title{tags}
\usage{
tags(object)

\S4method{tags}{groups}(object)

\S4method{tags}{tags}(object)
}
\arguments{
\item{object}{object from which tag snps are to be extracted}
}
\value{
character vector of tag SNPs
}
\description{
Accessor function to show tags from an object of class groups
}
\author{
Chris Wallace
}

