% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-snpmatrix.R
\name{textvenn}
\alias{textvenn}
\title{textvenn}
\usage{
textvenn(A, B, quiet = FALSE)
}
\arguments{
\item{A}{a vector}

\item{B}{a vector of the same type as A}

\item{quiet}{suppress printing if TRUE}
}
\value{
a list with three elements: A is setdiff(A,B).  int is intersect(A,B).  B is setdiff(B,A)
}
\description{
Utility function to calculate intersection and setdiffs for two vectors
}
\details{
Originally from randomFunctions
}
\examples{
A <- 1:3
B <- 2:4
textvenn(A,B)
tv <- textvenn(A,B,quiet=TRUE)
tv
}
\author{
Chris Wallace
}
