% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{union}
\alias{union}
\alias{union,snppicker,snppicker-method}
\alias{union,tags,tags-method}
\alias{union,groups,groups-method}
\title{Create a union of groups, snppicker or tags objects}
\usage{
union(x, y)

\S4method{union}{snppicker,snppicker}(x, y)

\S4method{union}{tags,tags}(x, y)

\S4method{union}{groups,groups}(x, y)
}
\arguments{
\item{x}{object of class \code{groups}, \code{snppicker} or \code{tags}}

\item{y}{object of same class as x}
}
\value{
object of class groups
}
\description{
First, objects are converted to class groups.  Then any groups
which overlap are identified, and replaced by their union.  Groups
which do not overlap are unchanged.  The combined set of groups is
returned.
}
\author{
Chris Wallace
}
