% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{xscale}
\alias{xscale}
\title{xscale}
\usage{
xscale(x, torange, xrange = c(min(x), max(x)))
}
\arguments{
\item{x}{numeric vector to be transformed}

\item{torange}{new range}

\item{xrange}{optional. full range of x vector if you are scaling only part of it. calculated from x if not supplied.}
}
\value{
transformed numeric vector
}
\description{
internal function: xscale
}
\details{
Linear transformation to map x to a new range, so min(x)==xrange[1], and max(x)==xrange[2]
}
\examples{
x<-1:10
cbind(x,
      GUESSFM:::xscale(x,c(0.1,1)),
      GUESSFM:::xscale(x,c(1319,20578)))
## Now use only part of x
x<-x[2:7]
## WRONG
cbind(x,
      GUESSFM:::xscale(x,c(0.1,1)),
      GUESSFM:::xscale(x,c(1319,20578)))
## RIGHT
cbind(x,
      GUESSFM:::xscale(x,c(0.1,1),xrange=c(1,10)),
      GUESSFM:::xscale(x,c(1319,20578),xrange=c(1,10)))
}
\author{
Chris Wallace
}
