% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{addlines}
\alias{addlines}
\title{addlines}
\usage{
addlines(summ, alpha = 0.5, size = 0.2, xvar = "position.plot")
}
\arguments{
\item{summ}{data.frame generated by \code{\link{guess.summ}()}}

\item{alpha}{transparency, default=0.5}

\item{size}{line width, default=0.2}

\item{xvar}{name of variable in summ which contains the x coordinates at which vertical lines should be plotted}
}
\value{
object that can be added to a ggplot
}
\description{
Add lines to a ggplot, indicating position of snps
}
\seealso{
Other plotting GUESSFM results: \code{\link{ggbed}},
  \code{\link{ggchr}}, \code{\link{ggld}},
  \code{\link{pp.nsnp}}, \code{\link{scalepos}},
  \code{\link{signal.plot}}
}
\author{
Chris Wallace
}
