% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{convert}
\alias{convert}
\alias{convert,groups-method}
\alias{convert,tags-method}
\title{Convert from old definitions of groups, tags classes to new}
\usage{
convert(object)

\S4method{convert}{groups}(object)

\S4method{convert}{tags}(object)
}
\arguments{
\item{object}{GUESSFM object from pre-development version}
}
\value{
new S4 structure
}
\description{
DON'T USE THIS FUNCTION UNLESS YOU HAVE OBJECTS STORED FROM A
PREVIOUS PRE-DEVELOPMENT VERSION!
}
