% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cummean}
\alias{cummean}
\title{internal function: cumulative mean}
\usage{
cummean(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
vector of cumulative mean
}
\description{
internal function: cumulative mean
}
\examples{
GUESSFM:::cummean(1:10)
}
