% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggbed}
\alias{ggbed}
\title{Plot a bed file}
\usage{
ggbed(bed, summ)
}
\arguments{
\item{bed}{data.frame representing data read in from a bed file}

\item{summ}{data.frame generated by \code{\link{guess.summ}()}}
}
\value{
ggplot object
}
\description{
Plot a bed file, overlaying lines showing the positions of SNPs in a summ object
}
\seealso{
Other plotting GUESSFM results: \code{\link{addlines}},
  \code{\link{ggchr}}, \code{\link{ggld}},
  \code{\link{pp.nsnp}}, \code{\link{scalepos}},
  \code{\link{signal.plot}}
}
\author{
chris
}
