% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggchr}
\alias{ggchr}
\title{ggchr}
\usage{
ggchr(summ)
}
\arguments{
\item{summ}{data.frame generated by \code{\link{guess.summ}()}}
}
\value{
ggplot2 object
}
\description{
Plot lines connecting each SNP's physical position to their representative position in the summ object
}
\seealso{
Other plotting GUESSFM results: \code{\link{addlines}},
  \code{\link{ggbed}}, \code{\link{ggld}},
  \code{\link{pp.nsnp}}, \code{\link{scalepos}},
  \code{\link{signal.plot}}
}
\author{
chris
}
