% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggld}
\alias{ggld}
\title{ggld}
\usage{
ggld(data, summ)
}
\arguments{
\item{data}{object of class SnpMatrix from which LD will be calculated}

\item{summ}{data.frame generated by \code{\link{guess.summ}()}}
}
\value{
ggplot output
}
\description{
Rotated LD plot
}
\details{
Generates a plot of r2, rotating through 45 degrees so it can be
aligned with other results, similar to Haploview.

The positions in the summ object are used to align the snps to
the correct coordinates.
}
\seealso{
Other plotting GUESSFM results: \code{\link{addlines}},
  \code{\link{ggbed}}, \code{\link{ggchr}},
  \code{\link{pp.nsnp}}, \code{\link{scalepos}},
  \code{\link{signal.plot}}
}
\author{
Chris Wallace
}
