% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{groups-class}
\alias{groups-class}
\alias{show,groups-method}
\alias{show,tags-method}
\alias{tags-class}
\title{groups class}
\usage{
\S4method{show}{tags}(object)

\S4method{show}{groups}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
groups class

tags class
}
\section{Slots}{

\describe{
\item{\code{tags}}{character vector giving tag SNPs.  Each tag indexes one group of SNPs}

\item{\code{.Data}}{list of character vectors giving the SNP membership of each group}

\item{\code{tags}}{character vector giving tag SNPs, one per SNP in \code{snps}, repeated as necessary}

\item{\code{.Data}}{character vector giving SNPs included in this tags object}
}}

