% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{groups-class}
\alias{groups-class}
\alias{tags-class}
\title{Group focused class for holding information about sets of SNPs
defined by their mutual LD}
\description{
\code{groups} and \code{tags} are two structures for holding the
same information, depending on whether your focus is on the sets of
SNPs or their index members.  It is easy to convert one to another
and perhaps, in future, one class may be deprecated.

\code{groups} and \code{tags} are two structures for holding the
same information, depending on whether your focus is on the sets of
SNPs or their index members.  It is easy to convert one to another
and perhaps, in future, one class may be deprecated.
}
\section{Slots}{

\describe{
\item{\code{tags}}{character vector giving tag SNPs.  Each tag indexes one group of SNPs}

\item{\code{.Data}}{list of character vectors giving the SNP membership of each group}

\item{\code{tags}}{character vector giving tag SNPs, one per SNP in
in \code{.Data}, repeated as necessary}

\item{\code{.Data}}{character vector giving SNPs included in this tags object}
}}

\examples{
new("groups")
new("tags")
}
