% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{tagsof}
\alias{tagsof}
\alias{taggedby}
\alias{taggedby,tags,character-method}
\alias{tagsof,tags,character-method}
\alias{taggedby,groups,character-method}
\alias{tagsof,groups,character-method}
\alias{[,groups,character,missing,missing-method}
\alias{[,groups,numeric,missing,missing-method}
\alias{[,groups,logical,missing,missing-method}
\alias{[,tags,character,missing,missing-method}
\alias{[[,groups,numeric-method}
\alias{[[,groups,logical-method}
\alias{[[,groups,character-method}
\title{\code{tagsof} shows tags for a named character vector of SNPs}
\usage{
tagsof(object, i)

taggedby(object, i)

\S4method{taggedby}{tags,character}(object, i)

\S4method{tagsof}{tags,character}(object, i)

\S4method{taggedby}{groups,character}(object, i)

\S4method{tagsof}{groups,character}(object, i)

\S4method{[}{groups,character,missing,missing}(x, i)

\S4method{[}{groups,numeric,missing,missing}(x, i)

\S4method{[}{groups,logical,missing,missing}(x, i)

\S4method{[}{tags,character,missing,missing}(x, i)

\S4method{[[}{groups,numeric}(x, i)

\S4method{[[}{groups,logical}(x, i)

\S4method{[[}{groups,character}(x, i)
}
\arguments{
\item{object}{tags or groups object}

\item{i}{numeric, logical or character vector to index SNPs or tags}

\item{x}{groups or tags object}
}
\value{
data.frame of tags and their tagged SNPs

subsetted groups or tags object
}
\description{
\code{tagsof} shows tags for a named character vector of SNPs

\code{taggedby} shows SNPs tagged by a named character vector of tag SNPs

'[' will extract another object of the same class.  '[[' will extract a single element.
}
