% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{guess.read}
\alias{guess.read}
\title{guess.read}
\usage{
guess.read(...)
}
\arguments{
\item{...}{arguments to be passed to read.snpmod}
}
\value{
see read.snpmod
}
\description{
guess.read, for backwards compatability
}
\details{
Just a wrapper for read.snpmod.
}
\author{
Chris Wallace
}

