% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{plot,snppicker,missing-method}
\alias{plot,snppicker,missing-method}
\alias{plot,ppnsnp,missing-method}
\title{Plots}
\usage{
\S4method{plot}{snppicker,missing}(x, do.plot = TRUE)

\S4method{plot}{ppnsnp,missing}(x)
}
\arguments{
\item{x}{ppsnp or snppicker object}

\item{do.plot}{if TRUE (default) print the plot on the current
device, otherwise invisibly return the plot object.}
}
\description{
Descriptive plots showing how the sets in snppicker objects were
generated.  Uses \code{ggplot2}, so you can customize using
\code{theme} etc.
}
