% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diffuse}
\alias{plot_diffuse}
\title{Diffusion plot}
\usage{
plot_diffuse(results, maxrank = 1000, thin = 500)
}
\arguments{
\item{results}{an object of class snpmod, or a named list of such objects}

\item{maxrank}{truncate the x axis at this value}

\item{thin}{thin data to display only \code{thin} points per line}
}
\value{
the ggplot2 object, which by default is print()ed to the current graphics device
}
\description{
A summary plot to indicate how diffuse the posterior is.
}
\details{
This is measured by how rapidly the cumulative posterior
probability increases with number of models.  For one dataset,
this may be informative in an absolute sense, but comparing
different datasets, by supplying a list of results, can be useful
too.

If a list of snpmod objects are given, multiple diffusion plots
are overlaid, with colours indicating labels.
}
\author{
Chris Wallace
}
