% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pp.nsnp}
\alias{pp.nsnp}
\title{Summarize the posterior model support by the number of SNPs contained in a model}
\usage{
pp.nsnp(results, plot = FALSE, expected = NULL, overdispersion = 1)
}
\arguments{
\item{results}{object of class snpmod}

\item{plot}{if TRUE, print a plot to current plotting device}

\item{expected}{optional, if specified, this is the prior expectation of the number of causal SNPs underlying the trait in the studied region.  Specifying will cause a line for the prior to be added to the plot}

\item{overdispersion}{optional, default=1.  Ignored unless expected is given.  The overdispersion of the beta binomial distribution used for the prior relative to a binomial.}
}
\value{
a named list containing
\itemize{
 \item{"pp"}{a named vector summarizing the posterior for each count of SNPs}
 \item{"plot"}{a ggplot object}
}
}
\description{
Summarize the posterior model support by the number of SNPs contained in a model
}
\seealso{
\link{snpprior}

Other plotting GUESSFM results: \code{\link{addlines}},
  \code{\link{ggbed}}, \code{\link{ggchr}},
  \code{\link{ggld}}, \code{\link{scalepos}},
  \code{\link{signal.plot}}
}
