% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read.snpmod}
\alias{read.snpmod}
\title{read.snpmod}
\usage{
read.snpmod(f, offset = 0)
}
\arguments{
\item{f}{filename}

\item{offset}{number of covariates}
}
\value{
object of class snpmod
}
\description{
Read in the output files and create an object of class snpmod
}
\author{
Chris Wallace
}
