% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{skewness}
\alias{skewness}
\title{skewness}
\usage{
skewness(x)
}
\arguments{
\item{x}{pp.nsnp object}
}
\value{
list of numerics
}
\description{
Summarise skewness of a posterior distribution on the number of SNPs in a model
}
\author{
Chris Wallace
}
