\docType{class}
\name{snppicker-class}
\alias{snppicker-class}
\title{Class to hold results of snp.picker algorithm}
\description{
\code{snp.picker} groups SNPs according to LD and model
inclusion and outputs objects of class snppicker.
}
\section{Slots}{

\describe{
\item{\code{groups}}{list of character vectors of SNPs in each group}

\item{\code{plotsdata}}{list of additional data relating to the snp.picker
process that allows a summary of that process to be plotted via
\code{plot}.}
}}
\section{Slots}{

}

