% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{summary,snppicker-method}
\alias{summary,snppicker-method}
\alias{summary,groups-method}
\title{Summaries}
\usage{
\S4method{summary}{snppicker}(object)

\S4method{summary}{groups}(object)
}
\arguments{
\item{object}{the thing to summarise}
}
\description{
Print summary of an object
}
