% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag}
\alias{tag}
\title{tag}
\usage{
tag(X, tag.threshold = 0.99, snps = NULL, samples = NULL, strata = NULL,
  quiet = FALSE, method = "single", split.at = 500)
}
\arguments{
\item{X}{SnpMatrix object}

\item{tag.threshold}{threshold to cut tree, default=0.99}

\item{snps}{colnames of the SnpMatrix object to be used}

\item{samples}{optional, subset of samples to use}

\item{strata}{optional, return a list of tag vectors, one for each stratum defined by as.factor(strata)}

\item{quiet}{if FALSE (default), show progress messages}

\item{method}{method used for heirarchical clustering.  See hclust for options.}

\item{split.at}{for large numbers of SNPs, tag first splits them into subsets of size split.at with similar MAF, then groups tag groups in high LD between subsets.  If you wish to avoid this, at additional computational cost, set split.at=NULL}
}
\value{
character vector, names are \code{snps}, values are the tag for each SNP
}
\description{
Derive tag SNPs for a SnpMatrix object using heirarchical clustering
}
\details{
Uses complete linkage and the \code{\link{hclust}} function to define clusters,
then cuts the tree at 1-tag.threshold
}
\author{
Chris Wallace
}
